requirements.txt
flask
google-generativeai
markdown
python-dotenv